+++ src/main/java/nom/bdezonia/zorbage/type/data/float16/real/Float16Member.java
@@ -469,7 +469,7 @@ public final class Float16Member
 		if (component < 0)
 			throw new IllegalArgumentException(
 					"negative component index error");
-		if (component == 0) return BigInteger.valueOf((long) v());
+		if (component == 0) return BigDecimal.valueOf(v()).toBigInteger();
 		return BigInteger.ZERO;
 	}
 
@@ -632,7 +632,7 @@ public final class Float16Member
 			return BigInteger.ZERO;
 		}
 		else {
-			return BigInteger.valueOf((long) v());
+			return BigDecimal.valueOf(v()).toBigInteger();
 		}
 	}
